Protobuf messages scenario, using core bazel proto rules
Necessitated revisiting how we determine what libraries are
not based on visible source files (and therefore, that should
be excluded).

initial proto example. triggered rethinking of how library dependencies for a module are determined.
proto example of a proto dependency/import situation. exclude all artifacts that are the products of source files visible in the current module, and in any parent modules.
scenario test for protobuf, ensuring expected module libraries are present (and no others)