drop a vcs.xml in place when creating the intellij project. precipitated rethink of iml / modules.xml orgnaization.
(the latter probably should have been does in a separate branch, oh well)

iml modules are now symlinked from within the source dirs, which greatly
simplifies the iml files and modules.xml, and overall makes this even closer
to a vanilla Intellij project layout.

stop making the iml content root be outside the MODULE_DIR. the content root is the module dir - this has the benefit of being much easier to reason about, and easy to troubleshoot iml problems. but, the challenge is now that these need to be symlinked down into the proper location in bazel output dirs, by the project setup script.
iml reorg
intellij_iml rule returns a script that installs it in the module dir - that is, that creates a symlink down into the bazel-generated iml file.
use a single relative path (not sure about this...the script is a little weirder.)
beginning of modules.xml revamp, to also create the iml symlinks
orient the modules.xml installer so that it expects to run at the workspace root
modules.xml bash does the right thing, i think
modules.xml refers to iml symlinks just under project root
fixes for scenario tests