Redo of how java sources, java_library, and intellij_iml relate, notably unifying iml and java_library
This makes it so we can cleanly determine what bazel artifacts are, and aren't,
generated by intellij source modules. We want to exclude all of these artifacts
from iml files (this is the code we're editing and compiling etc in the ide,
after all), and include all jar's that are not a result of compilation of these
sources.

This has also caused the iml parallel/redundant hierarchy (i.e., parallel
to the dependencies of various java_library targets) to collapse into a
single hierarchy: because now an iml target is a valid java dependency
in and of itself.

In sum, this set of changes gets rid of a major source of problems/weirdness,
and makes these rules much sounder and terser overall.

an intellij java source rule, that duck-types like a java rule, and so works with other java rules (as a dep). now we have a clean way to figure out what generated jars to exclude when putting together intellij module library lists. yay
extract higher level function, make source folders work
gather up transitive intellij deps so we know what libs to exclude
ability to depend on iml's as if they are java deps. iml's act like their 'compile' java_source dependency
the rules BUILD uses the new source lib style
start stripping out old attribs
use of new intellij_source_java_library across all scenarios. cleans up a lot of weirdness